Remove `CaptureSmithyConnectionWrapper`
