+++ rust-runtime/aws-smithy-http/src/connection.rs
@@ -7,7 +7,7 @@
 
 use std::fmt::{Debug, Formatter};
 use std::net::SocketAddr;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 
 /// Metadata that tracks the state of an active connection.
 #[derive(Clone)]
@@ -49,59 +49,5 @@ impl Debug for ConnectionMetadata {
             .field("is_proxied", &self.is_proxied)
             .field("remote_addr", &self.remote_addr)
             .finish()
-    }
-}
-
-type LoaderFn = dyn Fn() -> Option<ConnectionMetadata> + Send + Sync;
-
-/// State for a middleware that will monitor and manage connections.
-#[allow(missing_debug_implementations)]
-#[derive(Clone, Default)]
-pub struct CaptureSmithyConnection {
-    loader: Arc<Mutex<Option<Box<LoaderFn>>>>,
-}
-
-impl CaptureSmithyConnection {
-    /// Create a new connection monitor.
-    pub fn new() -> Self {
-        Self {
-            loader: Default::default(),
-        }
-    }
-
-    /// Set the retriever that will capture the `hyper` connection.
-    pub fn set_connection_retriever<F>(&self, f: F)
-    where
-        F: Fn() -> Option<ConnectionMetadata> + Send + Sync + 'static,
-    {
-        *self.loader.lock().unwrap() = Some(Box::new(f));
-    }
-
-    /// Get the associated connection metadata.
-    pub fn get(&self) -> Option<ConnectionMetadata> {
-        match self.loader.lock().unwrap().as_ref() {
-            Some(loader) => loader(),
-            None => {
-                tracing::debug!("no loader was set on the CaptureSmithyConnection");
-                None
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use crate::connection::{CaptureSmithyConnection, ConnectionMetadata};
-
-    #[test]
-    #[allow(clippy::redundant_clone)]
-    fn retrieve_connection_metadata() {
-        let retriever = CaptureSmithyConnection::new();
-        let retriever_clone = retriever.clone();
-        assert!(retriever.get().is_none());
-        retriever.set_connection_retriever(|| Some(ConnectionMetadata::new(true, None, || {})));
-
-        assert!(retriever.get().is_some());
-        assert!(retriever_clone.get().is_some());
     }
 }
